at86rf230: straighten out RX vs. TX synchronization
The is_tx flag did not protect adequately against races when switching
from receiving to transmitting. This patch removes the unnecessary
spinlock and enforces proper synchronization along the interrupt
processing chain (source, handler, and worker), before entering
transmit mode.

The mechanism for allowing transmits to be interrupted had two flaws:
1) it tried to return with the transmission potentially never
completed, and
2) it attempted to leave BUSY_TX with STATE_RX_ON, which according to
the data sheet is not valid.

This patch force the transceiver into PLL_ON/TX_ON state on
interruption and then has a common path for transiting to RX_ON and
clearing is_tx.

Note that the synchronization overhead introduced by this patch
is about 100 us on a Ben NanoNote.

Signed-off-by: Werner Almesberger <werner@almesberger.net>