ieee802154/cc2420: Check for FCS bit and skip frames with wrong CRC.
The cc2420 does offer hardware CRC calculation and we are using it in this
driver. Once the CRC is checked in hardware the two CRC bytes get replaced with
measurement informations and a FCS bit flag telling about the result of
the CRC check.

So far the driver did pass all frmaes into the stack without checking for the
correct CRC result. This patch checks if the FCS flag is good and drops the
frame if not.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>