ieee802154/cc2420: Fix hardware address recognition for ieee802154 long addresses
The ieee802154 long address was written into ram in the wrong order. The result
was that only frame with a given short address did pass the hardware address
recognition. This will already fail in the initial case when a node requests a
short address from the pan coordinator.

Also telling the stack that we are doing hardware acks to avoid warnings about
ack requested but not send frames.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>