at86rf230: add platform-specific reset function
This patch makes the setting of the "rstn" GPIO optional and
introduces a platform-provided reset function.

Such a reset function is necessary for boards that don't have a
direct GPIO that leads to the receiver or that have to follow a
special reset procedure, e.g., when reset is done by power-cycling.

Signed-off-by: Werner Almesberger <werner@almesberger.net>