ieee802154: remove ieee802154 policy from globals
This item is used inside ieee802154 layer only and it shouldn't be
defined in headers, common for multiple layers. But at the same time it
can't be marked as static because it's used in several files inside
ieee802154 layer. So just move ieee802154 netlink policy array
definition to more appropriate place.

Signed-off-by: Alexander Smirnov <alex.bluesman.smirnov@gmail.com>